assignment2/requirements.txt
svgpathtools==1.4.1
svgwrite==1.4.1
CairoSVG==2.5.2
Pillow==8.1.0
numpy==1.19.2
matplotlib==3.2.2
opencv-python==4.5.1.48
